server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.0
google-search-results==2.4.2
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-perplexity==0.1.1
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-cli==0.3.3
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
limits==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
numpy==2.2.6
openai==1.78.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjwt==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pyyaml==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.4
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
strenum==0.4.15
stripe==12.2.0
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
